Revert "sniffer is developed"
This reverts commit 8b3b5cffbcfeba0bcceedb7387b9adf00fe0bcaa.
